bug fixes for the cache

+++ ToolKit/Cache/TKAbstractStorage.h
@@ -1,5 +1,5 @@
 /**
- * @file    TKAbstractStorage.h
+ * @file    TKAbstractStorage.m
  * @author  Zsolt Mikola
  * @copyright Zsolt Mikola. All rights reserved.
  */
@@ -9,18 +9,18 @@
 @interface TKAbstractStorage : NSObject
 
 @property (nonatomic, strong) NSArray* sortDescriptors;
-@property (nonatomic, assign) int sizeLimit;
+@property (nonatomic, assign) NSUInteger sizeLimit;
 @property (nonatomic, assign) short countLimit;
-@property (nonatomic, assign) int hits;
-@property (nonatomic, assign) int misses;
-@property (nonatomic, assign) int size;
-@property (nonatomic, assign) int count;
+@property (nonatomic, assign) NSUInteger hits;
+@property (nonatomic, assign) NSUInteger misses;
+@property (nonatomic, assign) NSUInteger size;
+@property (nonatomic, assign) NSUInteger count;
 
 - (BOOL)insertObject:(id)object atIndex:(TKIndex*)index;
 - (id) objectAtIndex:(TKIndex*)index;
 - (BOOL)removeObjectAtIndex:(TKIndex*)index;
 
-- (void)reduceSizeTo:(int)size;
+- (void)reduceSizeTo:(NSUInteger)size;
 - (void)reduceCountTo:(short)count;
 - (void)rebuildIndex;
