+++ ToolKit/Cache/TKDiskStorage.m
@@ -134,7 +134,7 @@ static const short kSmartCacheDiskCountLimit = SHRT_MAX;
 
 - (NSString*)filepathWithIndex:(TKIndex*)index{
 
-    NSString* filename = [NSString stringWithFormat:@"%@-%d-%d-%d-%d", index.key, index.creationTime, index.expirationTime, index.priority, index.size];
+    NSString* filename = [NSString stringWithFormat:@"%@-%d-%d-%d-%lu", index.key, index.creationTime, index.expirationTime, index.priority, (unsigned long)index.size];
     return [_storagePath stringByAppendingPathComponent:filename];
 }
 
@@ -153,8 +153,8 @@ static const short kSmartCacheDiskCountLimit = SHRT_MAX;
 }
 
 - (NSString *)description{
-    return [NSString stringWithFormat:@"<%@: %p; size: %d; count: %d; hits: %d; misses: %d; sizeLimit: %d; countLimit: %d;>",
-            NSStringFromClass([self class]), self, self.size, self.count, self.hits, self.misses, self.sizeLimit, self.countLimit];
+    return [NSString stringWithFormat:@"<%@: %p; size: %lu; count: %lu; hits: %lu; misses: %lu; sizeLimit: %lu; countLimit: %d;>",
+            NSStringFromClass([self class]), self, (unsigned long)self.size, (unsigned long)self.count, (unsigned long)self.hits, (unsigned long)self.misses, (unsigned long)self.sizeLimit, self.countLimit];
 }
 
 @end
